refactor wargame and template mappers
