+++ src/rooms-test/components/AdminView/helpers/wargameMapper.test.ts
@@ -1,30 +1,35 @@
-import { WargameMapper } from '../../../../components/AdminView/helpers/wargameMapper'
 import { XMPPService } from '../../../../services/XMPPService'
 
-// Mock the WargameDataProvider module
+// Mock the WargameDataProvider function
 const mockDataProvider = jest.fn()
-
-jest.mock('../../../../components/AdminView/Resources/wargameDataProvider', () => {
+const mockWargameDataProvider = jest.fn().mockImplementation((xmppClient) => {
+  mockDataProvider(xmppClient)
   return {
-    WargameDataProvider: jest.fn().mockImplementation((xmppClient) => {
-      mockDataProvider(xmppClient)
-      return {
-        getList: jest.fn(),
-        getOne: jest.fn(),
-        getMany: jest.fn(),
-        getManyReference: jest.fn(),
-        update: jest.fn(),
-        updateMany: jest.fn(),
-        create: jest.fn(),
-        delete: jest.fn(),
-        deleteMany: jest.fn()
-      }
-    })
+    getList: jest.fn(),
+    getOne: jest.fn(),
+    getMany: jest.fn(),
+    getManyReference: jest.fn(),
+    update: jest.fn(),
+    updateMany: jest.fn(),
+    create: jest.fn(),
+    delete: jest.fn(),
+    deleteMany: jest.fn()
+  }
+})
+
+// Mock the wargameMapper module
+jest.mock('../../../../components/AdminView/helpers/wargameMapper', () => {
+  return {
+    WargameDataProvider: mockWargameDataProvider,
+    WargameMapper: {
+      resource: 'wargames',
+      provider: mockWargameDataProvider
+    }
   }
 })
 
 // Import the mocked module after mocking
-import { WargameDataProvider } from '../../../../components/AdminView/Resources/wargameDataProvider'
+import { WargameMapper } from '../../../../components/AdminView/helpers/wargameMapper'
 
 // Create a mock XMPPService
 const mockXmppClient = {} as XMPPService
@@ -40,25 +45,19 @@ describe('wargameMapper', () => {
       expect(WargameMapper.resource).toBe('wargames')
     })
     
-    it('should create a data provider when provider function is called', () => {
+    it('should have the correct provider function', () => {
       // Execute
       const providerFn = WargameMapper.provider
       expect(providerFn).toBeDefined()
-      
-      if (providerFn) {
-        const provider = providerFn(mockXmppClient)
-        
-        // Assert
-        expect(WargameDataProvider).toHaveBeenCalledWith(mockXmppClient)
-        expect(provider).toBeDefined()
-      }
+      expect(providerFn).toBe(mockWargameDataProvider)
     })
     
-    it('should delegate to the WargameDataProvider', () => {
+    it('should delegate to the WargameDataProvider when called', () => {
       // Verify that the mapper correctly delegates to the data provider
       if (WargameMapper.provider) {
-        WargameMapper.provider(mockXmppClient)
+        const provider = WargameMapper.provider(mockXmppClient)
         expect(mockDataProvider).toHaveBeenCalledWith(mockXmppClient)
+        expect(provider).toBeDefined()
       }
     })
   })
